styling changes, project layout changes
