+++ app/components/Animations/Typing/typing.module.css
@@ -2,7 +2,7 @@
     font-size: 10px;
     letter-spacing: 10px;
     text-transform: uppercase;
-    width: 100%;
+        
     /* animation: typing;
     animation-duration: 1.5s;
     animation-timing-function: steps(30, end);
