--- app/components/Menu/Dots.tsx
@@ -1,7 +1,0 @@
-const Dots = () => <>
-    <li><a>·</a></li>
-    <li style={{ fontWeight: "bolder" }}><a>·</a></li>
-    <li><a>·</a></li>
-</>
-
-export default Dots;
